% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms_matrix.R
\name{print.forms_matrix}
\alias{print.forms_matrix}
\title{Print method for a forms_matrix object}
\usage{
\method{print}{forms_matrix}(x, show_values = FALSE, ...)
}
\arguments{
\item{x}{The \code{forms_matrix} object to print.}

\item{show_values}{A logical value. If \code{FALSE} (default), prints enum names.
If \code{TRUE}, prints the underlying integer values.}

\item{...}{Additional arguments passed to print (not used here).}
}
\value{
Invisibly returns the original object \code{x}.
}
\description{
Controls how the 'forms_matrix' object is displayed in the console.
}
\examples{

print(forms_matrix_get(num_forms = 4))

}
