% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeomorphon-package.R
\docType{package}
\name{rgeomorphon-package}
\alias{rgeomorphon}
\alias{rgeomorphon-package}
\title{rgeomorphon: A Lightweight Implementation of the 'Geomorphon' Algorithm}
\description{
A fast and low-dependency implementation of the 'geomorphon' terrain form classification algorithm of Jasiewicz and Stepinski (2013) based largely on the 'GRASS GIS' 'r.geomorphon' module. This implementation employs a novel parallel algorithm, written in C++ and without external dependencies.
}
\keyword{internal}
