% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms_matrix.R
\name{forms_matrix}
\alias{forms_matrix}
\title{Create a \code{forms_matrix} object}
\usage{
forms_matrix(x, levels)
}
\arguments{
\item{x}{Integer. A 9x9  matrix.}

\item{levels}{Named integer vector. Map of integer values to their string
names. This is typically the output of \code{get_forms_grass_enum()}.}
}
\value{
An object of class \code{c("forms_matrix", "matrix", "array")}.
}
\description{
This constructor function wraps a 9x9 integer matrix and associates it
with a set of levels, creating a 'forms_matrix' object.
}
