% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms_matrix.R
\name{forms_matrix_get}
\alias{forms_matrix_get}
\title{Get a \code{forms_matrix} for Geomorphon Classification}
\usage{
forms_matrix_get(num_forms = 10, levels = get_forms_grass_enum())
}
\arguments{
\item{num_forms}{Integer. The number of forms to classify, one of \code{4}, \code{5},
\code{6}, or \code{10}.}

\item{levels}{Named integer with values between 0 and 10 corresponding to
form class labels. Default: \code{get_forms_grass_enum()}}
}
\value{
An object of class \code{forms_matrix}
}
\description{
Gets one of the internally defined forms matrices. A form matrix is defined
for the classic 10-form output (default) as well as three simplified classes:
4-form, 5-form, and 6-form.
}
