% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeomorphon-package.R
\docType{package}
\name{rgeomorphon-package}
\alias{rgeomorphon}
\alias{rgeomorphon-package}
\title{rgeomorphon: A Lightweight Implementation of the Geomorphon Algorithm}
\description{
A lightweight implementation of the geomorphon terrain form classification algorithm of Jasiewicz and Stepinski (2013) \doi{10.1016/j.geomorph.2012.11.005} based largely on the 'GRASS GIS' 'r.geomorphon' module. This implementation employs a novel algorithm written in C++ and 'RcppParallel'.
}
\seealso{
\code{\link[=geomorphons]{geomorphons()}}
}
\author{
\strong{Maintainer}: Andrew Brown \email{brown.andrewg@gmail.com} (\href{https://orcid.org/0000-0002-4565-533X}{ORCID})

}
\keyword{internal}
